move extension traits to ext module
Small refactoring to keep similar things together.
