+++ thorin/src/lib.rs
@@ -18,6 +18,7 @@ use crate::{
 };
 
 mod error;
+mod ext;
 mod index;
 mod package;
 mod relocate;
